Modularize the web server
