+++ sickbeard/server/web/home/__init__.py
@@ -1,0 +1,7 @@
+
+from sickbeard.server.web.home.base import Home
+from sickbeard.server.web.home.add_shows import HomeAddShows
+from sickbeard.server.web.home.change_log import HomeChangeLog
+from sickbeard.server.web.home.irc import HomeIRC
+from sickbeard.server.web.home.news import HomeNews
+from sickbeard.server.web.home.post_process import HomePostProcess
